Major Refactor: instruments submodule is base, camera reimplements
