+++ arguslib/camera/__init__.py
@@ -1,0 +1,2 @@
+from .video import Video
+from .locator import *
