+++ arguslib/camera/calibration.py
