+++ arguslib/camera/direct_camera.py
