+++ arguslib/camera/locator.py
