+++ arguslib/camera/undistorted_camera.py
