+++ arguslib/camera/video.py
