--- arguslib/video/__init__.py
@@ -1,2 +1,0 @@
-from .video import *
-from .locator import *
