+++ docs/api/arguslib.aircraft.rst
@@ -1,0 +1,29 @@
+arguslib.aircraft package
+=========================
+
+Submodules
+----------
+
+arguslib.aircraft.aircraft\_interface module
+--------------------------------------------
+
+.. automodule:: arguslib.aircraft.aircraft_interface
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.aircraft.fleet module
+------------------------------
+
+.. automodule:: arguslib.aircraft.fleet
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: arguslib.aircraft
+   :members:
+   :show-inheritance:
+   :undoc-members:
