+++ docs/api/arguslib.camera.rst
@@ -1,0 +1,77 @@
+arguslib.camera package
+=======================
+
+Submodules
+----------
+
+arguslib.camera.calibration module
+----------------------------------
+
+.. automodule:: arguslib.camera.calibration
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.camera module
+-----------------------------
+
+.. automodule:: arguslib.camera.camera
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.camera\_array module
+------------------------------------
+
+.. automodule:: arguslib.camera.camera_array
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.direct\_camera module
+-------------------------------------
+
+.. automodule:: arguslib.camera.direct_camera
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.locator module
+------------------------------
+
+.. automodule:: arguslib.camera.locator
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.undistorted\_camera module
+------------------------------------------
+
+.. automodule:: arguslib.camera.undistorted_camera
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.video module
+----------------------------
+
+.. automodule:: arguslib.camera.video
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.camera.video\_interface module
+---------------------------------------
+
+.. automodule:: arguslib.camera.video_interface
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: arguslib.camera
+   :members:
+   :show-inheritance:
+   :undoc-members:
