+++ docs/api/arguslib.instruments.rst
@@ -1,0 +1,21 @@
+arguslib.instruments package
+============================
+
+Submodules
+----------
+
+arguslib.instruments.instruments module
+---------------------------------------
+
+.. automodule:: arguslib.instruments.instruments
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: arguslib.instruments
+   :members:
+   :show-inheritance:
+   :undoc-members:
