+++ docs/api/arguslib.misc.rst
@@ -1,0 +1,53 @@
+arguslib.misc package
+=====================
+
+Submodules
+----------
+
+arguslib.misc.geo module
+------------------------
+
+.. automodule:: arguslib.misc.geo
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.misc.met module
+------------------------
+
+.. automodule:: arguslib.misc.met
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.misc.plotting module
+-----------------------------
+
+.. automodule:: arguslib.misc.plotting
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.misc.thermo module
+---------------------------
+
+.. automodule:: arguslib.misc.thermo
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.misc.times module
+--------------------------
+
+.. automodule:: arguslib.misc.times
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: arguslib.misc
+   :members:
+   :show-inheritance:
+   :undoc-members:
