+++ docs/api/arguslib.radar.rst
@@ -1,0 +1,45 @@
+arguslib.radar package
+======================
+
+Submodules
+----------
+
+arguslib.radar.locator module
+-----------------------------
+
+.. automodule:: arguslib.radar.locator
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.radar.radar module
+---------------------------
+
+.. automodule:: arguslib.radar.radar
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.radar.radar\_interface module
+--------------------------------------
+
+.. automodule:: arguslib.radar.radar_interface
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+arguslib.radar.radar\_overlay\_interface module
+-----------------------------------------------
+
+.. automodule:: arguslib.radar.radar_overlay_interface
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: arguslib.radar
+   :members:
+   :show-inheritance:
+   :undoc-members:
