+++ docs/api/arguslib.rst
@@ -1,0 +1,22 @@
+arguslib package
+================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   arguslib.aircraft
+   arguslib.camera
+   arguslib.instruments
+   arguslib.misc
+   arguslib.radar
+
+Module contents
+---------------
+
+.. automodule:: arguslib
+   :members:
+   :show-inheritance:
+   :undoc-members:
