+++ docs/api/modules.rst
@@ -1,0 +1,7 @@
+arguslib
+========
+
+.. toctree::
+   :maxdepth: 4
+
+   arguslib
