+++ examples/quick_quicklooks.py
@@ -4,15 +4,15 @@
 import datetime
 from pathlib import Path
 
-from arguslib.instruments.camera import Camera
+from arguslib.camera.camera import Camera
 
 from arguslib.aircraft import AircraftInterface
 from arguslib.misc.geo import ft_to_km
 from arguslib.radar import RadarInterface
-from arguslib.instruments.radar import Radar
-from arguslib.instruments.camera_array import CameraArray
+from arguslib.radar.radar import Radar
+from arguslib.camera.camera_array import CameraArray
 
-from arguslib.instruments.undistorted_camera import UndistortedCamera
+from arguslib.camera.undistorted_camera import UndistortedCamera
 
 adsb_datadir = Path("/disk1/Data/ADS-B/COBALT/")
 outdir = Path(__file__).parent / "output" / "camera_radar_quicklooks"
